[HOL-GettingStartedWindowsAzureStorage] * Updated lab to VS 2013, MVC5 and latest Azure SDK * Reviewed flow of Exercise 4 and 5 (more streamlined now)

+++ HOLs/HOL-GettingStartedWindowsAzureStorage/Source/Ex3-UnderstandingStorageAbstractions/End/PhotoUploader_WebRole/Controllers/HomeController.cs
@@ -17,18 +17,13 @@ namespace PhotoUploader_WebRole.Controllers
     {
         private CloudStorageAccount StorageAccount = CloudStorageAccount.Parse(CloudConfigurationManager.GetSetting("StorageConnectionString"));
 
-        //
-        // GET: /
-
         public ActionResult Index()
         {
             var photoContext = this.GetPhotoContext();
             var photos = photoContext.GetPhotos();
-            return this.View(photos.Select(this.ToViewModel).ToList());
+            var photosViewModels = photos.Select(this.ToViewModel).ToList();
+            return this.View(photosViewModels);
         }
-
-        //
-        // GET: /Home/Details/5
 
         public async Task<ActionResult> Details(string partitionKey, string rowKey)
         {
@@ -49,16 +44,10 @@ namespace PhotoUploader_WebRole.Controllers
             return this.View(viewModel);
         }
 
-        //
-        // GET: /Home/Create
-
         public ActionResult Create()
         {
             return View();
         }
-
-        //
-        // POST: /Home/Create
 
         [HttpPost]
         public async Task<ActionResult> Create(PhotoViewModel photoViewModel, HttpPostedFileBase file, FormCollection collection)
@@ -88,22 +77,12 @@ namespace PhotoUploader_WebRole.Controllers
             var photoContext = this.GetPhotoContext();
             await photoContext.AddPhotoAsync(photo);
 
-            //Send create notification
-            try
-            {
-                var msg = new CloudQueueMessage("Photo Uploaded");
-                await this.GetCloudQueue().AddMessageAsync(msg);
-            }
-            catch (Exception e)
-            {
-                System.Diagnostics.Trace.TraceInformation("Error", "Couldn't send notification");
-            }
+            // Send create notification
+            var msg = new CloudQueueMessage("Photo Uploaded");
+            await this.GetCloudQueue().AddMessageAsync(msg);
 
             return this.RedirectToAction("Index");
         }
-
-        //
-        // GET: /Home/Edit/5
 
         public async Task<ActionResult> Edit(string partitionKey, string rowKey)
         {
@@ -124,9 +103,6 @@ namespace PhotoUploader_WebRole.Controllers
             return this.View(viewModel);
         }
 
-        //
-        // POST: /Home/Edit/5
-
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Edit(PhotoViewModel photoViewModel, FormCollection collection)
@@ -136,17 +112,21 @@ namespace PhotoUploader_WebRole.Controllers
                 return this.View();
             }
 
-            var photo = this.FromViewModel(photoViewModel);
-
-            //Update information in Table Storage
             var photoContext = this.GetPhotoContext();
-            await photoContext.UpdatePhotoAsync(photo);
+            var entityToUpdate = await photoContext.GetByIdAsync(photoViewModel.PartitionKey, photoViewModel.RowKey);
+
+            if (entityToUpdate == null)
+            {
+                return this.HttpNotFound();
+            }
+
+            // Update entity information from ViewModel
+            entityToUpdate.Title = photoViewModel.Title;
+            entityToUpdate.Description = photoViewModel.Description;
+            await photoContext.UpdatePhotoAsync(entityToUpdate);
 
             return this.RedirectToAction("Index");
         }
-
-        //
-        // GET: /Home/Delete/5
 
         public async Task<ActionResult> Delete(string partitionKey, string rowKey)
         {
@@ -167,15 +147,18 @@ namespace PhotoUploader_WebRole.Controllers
             return this.View(viewModel);
         }
 
-        //
-        // POST: /Home/Delete/5
-
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> DeleteConfirmed(string partitionKey, string rowKey)
         {
             var photoContext = this.GetPhotoContext();
             var photo = await photoContext.GetByIdAsync(partitionKey, rowKey);
+
+            if (photo == null)
+            {
+                return this.HttpNotFound();
+            }
+
             await photoContext.DeletePhotoAsync(photo);
 
             //Deletes the Image from Blob Storage
@@ -185,16 +168,9 @@ namespace PhotoUploader_WebRole.Controllers
                 await blob.DeleteIfExistsAsync();
             }
 
-            //Send delete notification
-            try
-            {
-                var msg = new CloudQueueMessage("Photo Deleted");
-                await this.GetCloudQueue().AddMessageAsync(msg);
-            }
-            catch (Exception e)
-            {
-                System.Diagnostics.Trace.TraceInformation("Error", "Couldn't send notification");
-            }
+            // Send delete notification
+            var msg = new CloudQueueMessage("Photo Deleted");
+            await this.GetCloudQueue().AddMessageAsync(msg);
 
             return this.RedirectToAction("Index");
         }
@@ -228,13 +204,8 @@ namespace PhotoUploader_WebRole.Controllers
 
         private CloudBlobContainer GetBlobContainer()
         {
-            var client = this.StorageAccount.CreateCloudBlobClient();
-            var container = client.GetContainerReference(CloudConfigurationManager.GetSetting("ContainerName"));
-            if (container.CreateIfNotExists())
-            {
-                container.SetPermissions(new BlobContainerPermissions { PublicAccess = BlobContainerPublicAccessType.Blob });
-            }
-
+            var cloudBlobClient = this.StorageAccount.CreateCloudBlobClient();
+            var container = cloudBlobClient.GetContainerReference(CloudConfigurationManager.GetSetting("ContainerName"));
             return container;
         }
 
@@ -242,7 +213,6 @@ namespace PhotoUploader_WebRole.Controllers
         {
             var queueClient = this.StorageAccount.CreateCloudQueueClient();
             var queue = queueClient.GetQueueReference("messagequeue");
-            queue.CreateIfNotExists();
             return queue;
         }
     }
