+++ HOLs/HOL-GettingStartedWindowsAzureStorage/Source/Ex5-UpdatingSASToUseStoredAccessPolicies/End/QueueProcessor_WorkerRole/WorkerRole.cs
@@ -9,9 +9,9 @@ using Microsoft.WindowsAzure.Diagnostics;
 using Microsoft.WindowsAzure.ServiceRuntime;
 using Microsoft.WindowsAzure.Storage;
 using Microsoft.WindowsAzure.Storage.Queue;
+using Microsoft.WindowsAzure.Storage.Queue.Protocol;
 using Microsoft.WindowsAzure.Storage.Auth;
 using Microsoft.WindowsAzure.Storage.Blob;
-using Microsoft.WindowsAzure.Storage.Queue.Protocol;
 
 namespace QueueProcessor_WorkerRole
 {
@@ -37,43 +37,46 @@ namespace QueueProcessor_WorkerRole
             {
                 Thread.Sleep(10000);
                 Trace.TraceInformation("Working", "Information");
-                if (DateTime.UtcNow.AddMinutes(1) >= this.serviceQueueSasExpiryTime)
+                if (queue.Exists())
                 {
-                    queueClient = new CloudQueueClient(this.uri, new StorageCredentials(this.GetProcessSasForQueues()));
-                    queue = queueClient.GetQueueReference("messagequeue");
-                }
-
-                var msg = queue.GetMessage();
-
-                if (msg != null)
-                {
-                    queue.FetchAttributes();
-
-                    var messageParts = msg.AsString.Split(new char[] { ',' });
-                    var message = messageParts[0];
-                    var blobReference = messageParts[1];
-
-                    if (queue.Metadata.ContainsKey("Resize") && string.Equals(message, "Photo Uploaded"))
+                    if (DateTime.UtcNow.AddMinutes(1) >= this.serviceQueueSasExpiryTime)
                     {
-                        var maxSize = queue.Metadata["Resize"];
-
-                        Trace.TraceInformation("Resize is configured");
-
-                        CloudBlockBlob outputBlob = this.container.GetBlockBlobReference(blobReference);
-
-                        outputBlob.FetchAttributes();
-
-                        Trace.TraceInformation(string.Format("Image ContentType: {0}", outputBlob.Properties.ContentType));
-                        Trace.TraceInformation(string.Format("Image width: {0}", outputBlob.Metadata["Width"]));
-                        Trace.TraceInformation(string.Format("Image hieght: {0}", outputBlob.Metadata["Height"]));
+                        queueClient = new CloudQueueClient(this.uri, new StorageCredentials(this.GetProcessSasForQueues()));
+                        queue = queueClient.GetQueueReference("messagequeue");
                     }
 
-                    Trace.TraceInformation(string.Format("Message '{0}' processed.", msg.AsString));
-                    queue.DeleteMessage(msg);
+                    var msg = queue.GetMessage();
 
+                    if (msg != null)
+                    {
+                        queue.FetchAttributes();
+
+                        var messageParts = msg.AsString.Split(new char[] { ',' });
+                        var message = messageParts[0];
+                        var blobReference = messageParts[1];
+
+                        if (queue.Metadata.ContainsKey("Resize") && string.Equals(message, "Photo Uploaded"))
+                        {
+                            var maxSize = queue.Metadata["Resize"];
+
+                            Trace.TraceInformation("Resize is configured");
+
+                            CloudBlockBlob outputBlob = this.container.GetBlockBlobReference(blobReference);
+
+                            outputBlob.FetchAttributes();
+
+                            Trace.TraceInformation(string.Format("Image ContentType: {0}", outputBlob.Properties.ContentType));
+                            Trace.TraceInformation(string.Format("Image width: {0}", outputBlob.Metadata["Width"]));
+                            Trace.TraceInformation(string.Format("Image hieght: {0}", outputBlob.Metadata["Height"]));
+                        }
+
+                        Trace.TraceInformation(string.Format("Message '{0}' processed.", msg.AsString));
+                        queue.DeleteMessage(msg);
+                    }
                 }
             }
         }
+
 
         public override bool OnStart()
         {
@@ -82,6 +85,7 @@ namespace QueueProcessor_WorkerRole
 
             // For information on handling configuration changes
             // see the MSDN topic at http://go.microsoft.com/fwlink/?LinkId=166357.
+
             this.CreateCloudBlobClient();
 
             return base.OnStart();
@@ -99,8 +103,8 @@ namespace QueueProcessor_WorkerRole
             queue.SetPermissions(qp);
 
             var token = queue.GetSharedAccessSignature(
-                             new SharedAccessQueuePolicy(),
-                                 "process");
+                            new SharedAccessQueuePolicy(),
+                            "process");
             this.serviceQueueSasExpiryTime = DateTime.UtcNow.AddMinutes(15);
             return token;
         }
